+++ Cargo.toml
@@ -47,6 +47,6 @@ serde_json = "1.0.122"
 clap = { version = "4.5.13", features = ["derive", "env"] }
 
 # arweave uploader
-# TODO: there are many unused stuff here, but everything breaks if you use the minimal set
+# NOTE: there are many unused stuff here, but everything breaks if you use the minimal set
 # because Bundlr SDK is not maintained at all
 bundlr-sdk = { version = "0.5.0" }
