+++ src/cli/commands/coordinator/view.rs
@@ -1,0 +1,89 @@
+use crate::{
+    contracts::{bytes_to_string, TaskStatus},
+    DriaOracle,
+};
+use alloy::{eips::BlockNumberOrTag, primitives::U256};
+use eyre::Result;
+
+impl DriaOracle {
+    /// Views the task events between two blocks, logs everything on screen.
+    pub(in crate::cli) async fn view_task_events(
+        &self,
+        from_block: impl Into<BlockNumberOrTag> + Clone,
+        to_block: impl Into<BlockNumberOrTag> + Clone,
+    ) -> Result<()> {
+        let from_block: BlockNumberOrTag = from_block.clone().into();
+        let to_block: BlockNumberOrTag = to_block.clone().into();
+        log::info!(
+            "Viewing task ids & statuses between blocks: {} - {}",
+            from_block
+                .as_number()
+                .map(|n| n.to_string())
+                .unwrap_or(from_block.to_string()),
+            to_block
+                .as_number()
+                .map(|n| n.to_string())
+                .unwrap_or(to_block.to_string())
+        );
+
+        let task_events = self.get_tasks_in_range(from_block, to_block).await?;
+
+        for (event, log) in task_events {
+            log::info!(
+                "Task {} changed from {} to {} at block {}",
+                event.taskId,
+                TaskStatus::try_from(event.statusBefore).unwrap_or_default(),
+                TaskStatus::try_from(event.statusAfter).unwrap_or_default(),
+                log.block_number.unwrap_or_default()
+            );
+        }
+
+        Ok(())
+    }
+
+    /// Views the request, responses and validations of a single task, logs everything on screen.
+    pub(in crate::cli) async fn view_task(&self, task_id: U256) -> Result<()> {
+        log::info!("Viewing task {}.", task_id);
+        let (request, responses, validations) = self.get_task(task_id).await?;
+
+        log::info!(
+            "Request Information:\nRequester: {}\nStatus:    {}\nInput:     {}\nModels:    {}",
+            request.requester,
+            TaskStatus::try_from(request.status)?,
+            bytes_to_string(&request.input)?,
+            bytes_to_string(&request.models)?
+        );
+
+        log::info!("Responses:");
+        if responses._0.is_empty() {
+            log::warn!("There are no responses yet.");
+        } else {
+            for (idx, response) in responses._0.iter().enumerate() {
+                log::info!(
+                    "Response  #{}\nOutput:    {}\nMetadata:  {}\nGenerator: {}",
+                    idx,
+                    bytes_to_string(&response.output)?,
+                    bytes_to_string(&response.metadata)?,
+                    response.responder
+                );
+            }
+        }
+
+        log::info!("Validations:");
+        if validations._0.is_empty() {
+            log::warn!("There are no validations yet.");
+        } else {
+            for (idx, validation) in validations._0.iter().enumerate() {
+                log::info!(
+                    "Validation #{}\nScores:     {:?}\nMetadata:   {}\nValidator:  {}",
+                    idx,
+                    validation.scores,
+                    bytes_to_string(&validation.metadata)?,
+                    validation.validator
+                );
+            }
+        }
+
+        Ok(())
+    }
+}
