+++ src/configurations/mod.rs
@@ -17,12 +17,6 @@ pub struct DriaOracleConfig {
     pub tx_timeout: Option<std::time::Duration>,
 }
 
-impl Default for DriaOracleConfig {
-    fn default() -> Self {
-        Self::new_from_env().unwrap().enable_logs()
-    }
-}
-
 impl DriaOracleConfig {
     pub fn new(secret_key: &B256, rpc_url: Url) -> Result<Self> {
         let signer =
@@ -50,8 +44,6 @@ impl DriaOracleConfig {
     /// Required environment variables:
     /// - `SECRET_KEY`
     /// - `RPC_URL`
-    /// - `TX_TIMEOUT_SECS` (optional)
-    #[deprecated = "do this in tests instead"]
     pub fn new_from_env() -> Result<Self> {
         // parse private key
         let private_key_hex = env::var("SECRET_KEY").wrap_err("SECRET_KEY is not set")?;
@@ -87,15 +79,6 @@ impl DriaOracleConfig {
     /// Change the underlying wallet.
     pub fn with_wallet(mut self, wallet: EthereumWallet) -> Self {
         self.wallet = wallet;
-        self
-    }
-
-    /// Enables `env_logger`.
-    #[deprecated = "do this in tests instead"]
-    pub fn enable_logs(self) -> Self {
-        if let Err(e) = env_logger::try_init() {
-            log::error!("Error during env_logger::try_init: {}", e);
-        }
         self
     }
